+++ interceptors/session.js
@@ -13,29 +13,29 @@
 
 const {resolve} = require("path");
 const tools = resolve(__dirname, "..", "tools");
-const [security, key, redisService, utilities, errors] = [require(resolve(tools, "security")), "cynomys-sig", require(resolve(__dirname, "..", "services", "redisService")),
+const [{decipher, cipher}, key, redisService, {randomCode}, errors] = [require(resolve(tools, "security")), "cynomys-portal", require(resolve(__dirname, "..", "services", "redisService")),
   require(resolve(tools, "utilities")), require(resolve(__dirname, "..", "errors"))];
 
 const parseLanguage = header => {
   let {language = header["accept-language"]} = header;
   try {
     language = language.split(";")[0];
-    let value = language.split(",");
+    const value = language.split(",");
     language = value.find(v => "en-US" === v);
     if (!language) {
       throw new Error("");
     }
-  } catch (error) {
+  } catch (err) {
     language = "zh-CN";
   }
   return language;
 };
 
 module.exports = function* (next) {
-  let [start, cookie, requestIp = "0.0.0.0", header] = [new Date(), this.cookies.get(key), this.request.ip, this.request.header];
+  let [start, cookie, requestIp = "0.0.0.0", header] = [Date.now(), this.cookies.get(key), this.request.ip, this.request.header];
   let {token} = header;
   this.requestIp = requestIp; // 绑定请求Ip到 this对象上
-  this.language = parseLanguage(header); // 绑定语言
+  this.language = parseLanguage(header); // 绑定请求语言
   if (!cookie) {
     cookie = token;
   }
@@ -62,56 +62,42 @@ module.exports = function* (next) {
     this.cookies.set(key, requestIp);
   };
 
-  this.getSession = function* (__token) {
+  this.getSession = function* (token) {
     try {
-			if(__token) {
-        cookie = __token;
+			if(token) {
+        cookie = token;
       }
-      let [secret, equipmentId, userId] = security.decipher(cookie).split(":"); // 解构 设备id,用户id,ip
-      if (equipmentId !== this.equipmentId) {
-        errors.throwSessionError(this.language);
-      }
-      let session = yield redisService.get(`${equipmentId}:${userId}`);
-      if (session.secret !== secret) {
-        errors.throwSessionError(this.language);
-      }
-      delete session.secret;
-      session.cookie = cookie;
+      const [permit, to] = security.decipher(cookie).split(":"); // 解构 设备id,用户id,ip
+      let session = yield redisService.get(`${to}-${permit}`);
       return session;
     } catch (error) {
       return null;
     }
   };
 
-  this.reSetSession = function* (value) {
-    let [secret, equipmentId, userId] = security.decipher(cookie).split(":"); // 解构 设备id,用户id,ip
-    if (equipmentId !== this.equipmentId) {
-      errors.throwSessionError(this.language);
-    }
-    let session = yield redisService.get(`${equipmentId}:${userId}`);
-    if (session.secret !== secret) {
-      errors.throwSessionError(this.language);
-    }
+  this.upDateSession = function* (value) {
+    const [permit, to] = security.decipher(cookie).split(":"); // 解构 设备id,用户id,ip
+    const _ = yield redisService.get(`${to}-${permit}`);
     Object.assign(session, value);
-    return yield redisService.set(`${equipmentId}:${userId}`, session);
+    return yield redisService.set(`${to}-${permit}`, session);
   };
 
   this.setSession = function* (seems, token = cookie) {
-    let [secret, equipmentId, userId] = security.decipher(token).split(":");
-    if (equipmentId !== this.equipmentId) {
-      errors.throwSessionError(this.language);
-    }
-    seems.secret = secret;
-    return yield redisService.set(`${equipmentId}:${userId}`, seems);
+    const [permit, to] = security.decipher(token).split(":");
+    Object.assign(seems, {
+      secret: permit,
+      timestamp: Date.now(),
+    });
+    return yield redisService.set(`${to}-${permit}`, seems);
   };
 
-  this.sign = function (userId) {
-    cookie = security.cipher(`${utilities.randomCode(15)}:${this.equipmentId}:${userId}`);
+  this.sign = (to, permit) => {
+    cookie = security.cipher(`${permit}:${to}`);
     this.cookies.set(key, cookie);
     return cookie;
   };
 
   yield next;
 
-  console.log("[%s][%s] %s %s - %s ms", new Date().toLocaleString(), requestIp, this.method, this.url, new Date() - start);
+  console.log("[%s][%s] %s %s - %s ms", new Date().toLocaleString(), requestIp, this.method, this.url, Date.now() - start);
 };